requirements.txt
asgiref==3.7.2
Django==4.2.4
django-rest-framework==0.1.0
djangorestframework==3.14.0
Pillow==10.0.0
psycopg2-binary==2.9.7
python-decouple==3.8
pytz==2023.3
sqlparse==0.4.4